+++ handler/connectgrpc.go
@@ -29,7 +29,6 @@ import (
 	sf "github.com/gogufo/gufo-api-gateway/gufodao"
 	pb "github.com/gogufo/gufo-api-gateway/proto/go"
 	"github.com/spf13/viper"
-	"google.golang.org/grpc"
 	pbv "gopkg.in/cheggaaa/pb.v1"
 )
 
@@ -44,97 +43,9 @@ type uploader struct {
 	FailRequest chan string
 }
 
-type PBRequest struct {
-	*pb.Request
-}
+func GetHostAndPort(t *pb.Request) (host string, port string, plygintype string) {
 
-func (t *PBRequest) MSCommunication(host string, port string) (answer map[string]interface{}) {
-
-	answer = make(map[string]interface{})
-
-	connection := fmt.Sprintf("%s:%s", host, port)
-
-	opts := []grpc.DialOption{
-		grpc.WithInsecure(),
-	}
-	//	args := os.Args
-	conn, err := grpc.Dial(connection, opts...)
-
-	if err != nil {
-
-		if viper.GetBool("server.sentry") {
-			sentry.CaptureException(err)
-		} else {
-			sf.SetErrorLog("connectgrpc: " + err.Error())
-		}
-
-		answer["httpcode"] = 400
-		answer["code"] = "0000234"
-		answer["message"] = err.Error()
-		return answer
-
-	}
-
-	defer conn.Close()
-
-	client := pb.NewReverseClient(conn)
-
-	response, err := client.Do(context.Background(), t.Request)
-
-	if err != nil {
-		if viper.GetBool("server.sentry") {
-			sentry.CaptureException(err)
-		} else {
-			sf.SetErrorLog("connectgrpc: " + err.Error())
-		}
-		answer["httpcode"] = 500
-		answer["code"] = "0000236"
-		answer["message"] = fmt.Sprintf("Module connection error: %s", err.Error())
-		return answer
-
-	}
-
-	answer = sf.ToMapStringInterface(response.Data)
-
-	//update *sf.Request
-	if response.RequestBack.Token != t.Token {
-		t.Token = response.RequestBack.Token
-	}
-	if response.RequestBack.TokenType != t.TokenType {
-		t.TokenType = response.RequestBack.TokenType
-	}
-	if response.RequestBack.Language != t.Language {
-		t.Language = response.RequestBack.Language
-	}
-	if response.RequestBack.UID != t.UID {
-		t.UID = response.RequestBack.UID
-	}
-	if response.RequestBack.IsAdmin != t.IsAdmin {
-		t.IsAdmin = response.RequestBack.IsAdmin
-	}
-	if response.RequestBack.SessionEnd != t.SessionEnd {
-		t.SessionEnd = response.RequestBack.SessionEnd
-	}
-	if response.RequestBack.Completed != t.Completed {
-		t.Completed = response.RequestBack.Completed
-	}
-	if response.RequestBack.Readonly != t.Readonly {
-		t.Readonly = response.RequestBack.Readonly
-	}
-
-	return answer
-
-}
-
-func connectgrpc(w http.ResponseWriter, r *http.Request, t *pb.Request) {
-
-	st := PBRequest{}
-	st.Request = t
-
-	port := ""
-	host := ""
 	pluginname := fmt.Sprintf("microservices.%s", *t.Module)
-	plygintype := ""
 
 	msmethod := viper.GetBool("server.masterservice")
 
@@ -144,25 +55,29 @@ func connectgrpc(w http.ResponseWriter, r *http.Request, t *pb.Request) {
 		host = viper.GetString("microservices.masterservice.host")
 		port = viper.GetString("microservices.masterservice.port")
 
-		//Save curent data
-		//	curparam := *t.Param
-		//	curmethod := *t.Method
-
-		//api/v3/auth/signup
-		//api/v3/masterservice/getmicroservicebypath
-
 		//Modify data for request masterservice
-		*st.Request.MS.Param = "getmicroservicebypath"
-		*st.Request.MS.Method = "GET"
+		mst := &pb.MasterService{}
+		param := "getmicroservicebypath"
+		gt := "GET"
+		mst.Param = &param
+		mst.Method = &gt
 
-		ans := st.MSCommunication(host, port)
+		t.MS = mst
+
+		ans := sf.GRPCConnect(host, port, t)
 		if ans["httpcode"] != nil {
-			httpcode := 0
 
-			httpcode, _ = strconv.Atoi(fmt.Sprintf("%v", ans["httpcode"]))
+			if viper.GetBool("server.sentry") {
+				sentry.CaptureMessage(fmt.Sprintf("%v", ans["message"]))
+			} else {
+				sf.SetErrorLog(fmt.Sprintf("%v", ans["message"]))
+			}
+			//	httpcode := 0
 
-			errorAnswer(w, r, t, httpcode, fmt.Sprintf("%v", ans["code"]), fmt.Sprintf("%v", ans["message"]))
-			return
+			//	httpcode, _ = strconv.Atoi(fmt.Sprintf("%v", ans["httpcode"]))
+
+			//	errorAnswer(w, r, t, httpcode, fmt.Sprintf("%v", ans["code"]), fmt.Sprintf("%v", ans["message"]))
+			return "", "", ""
 		}
 
 		host = fmt.Sprintf("%v", ans["host"])
@@ -187,8 +102,8 @@ func connectgrpc(w http.ResponseWriter, r *http.Request, t *pb.Request) {
 			} else {
 				sf.SetErrorLog(msg)
 			}
-			errorAnswer(w, r, t, 401, "0000235", msg)
-			return
+			//	errorAnswer(w, r, t, 401, "0000235", msg)
+			return "", "", ""
 		}
 
 		hostpath := fmt.Sprintf("%s.host", pluginname)
@@ -200,6 +115,13 @@ func connectgrpc(w http.ResponseWriter, r *http.Request, t *pb.Request) {
 	if !msmethod {
 		plygintype = fmt.Sprintf("%s.type", pluginname)
 	}
+
+	return host, port, plygintype
+}
+
+func connectgrpc(w http.ResponseWriter, r *http.Request, t *pb.Request) {
+
+	host, port, plygintype := GetHostAndPort(t)
 
 	if plygintype == "internal" {
 
@@ -218,8 +140,7 @@ func connectgrpc(w http.ResponseWriter, r *http.Request, t *pb.Request) {
 
 	}
 
-	ans := st.MSCommunication(host, port)
-	t = st.Request
+	ans := sf.GRPCConnect(host, port, t)
 
 	moduleAnswerv3(w, r, ans, t)